refactor(semantic): split cfg module
